refactor: add to defualt error handler - about execute result - move to unuseful file architecture
